requirements.txt
python-telegram-bot==13.15
pip>=25.3